update cost util

+++ app/utils/trades.py
@@ -3,4 +3,4 @@ import decimal
 def calculate_trade(shares, price, fee):
     cost = decimal.Decimal(shares*price)
     cost += fee
-    return float(round(cost, 2))
+    return round(cost, 2)
